+++ examples/pages/Styled/SingleSelect.js
@@ -1,0 +1,58 @@
+import React from 'react';
+import { withValue } from 'react-value';
+import { colourOptions } from '../../data';
+import Select from '../../../src';
+import chroma from 'chroma-js';
+
+const SelectWithValue = withValue(Select);
+
+const dot = (color = '#ccc') => ({
+  alignItems: 'center',
+  display: 'flex ',
+
+  ':before': {
+    backgroundColor: color,
+    borderRadius: 10,
+    content: ' ',
+    display: 'block',
+    marginRight: 8,
+    height: 10,
+    width: 10,
+  },
+});
+
+const colourStyles = {
+  control: styles => ({ ...styles, backgroundColor: 'white' }),
+  option: (styles, { data, isDisabled, isFocused, isSelected }) => {
+    const color = chroma(data.color);
+    return {
+      ...styles,
+      backgroundColor: isDisabled
+        ? null
+        : isSelected ? data.color : isFocused ? color.alpha(0.1).css() : null,
+      color: isDisabled
+        ? '#ccc'
+        : isSelected
+          ? chroma.contrast(color, 'white') > 2 ? 'white' : 'black'
+          : data.color,
+      cursor: isDisabled ? 'not-allowed' : 'default',
+    };
+  },
+};
+
+const colourStylesSingle = {
+  ...colourStyles,
+  input: styles => ({ ...styles, ...dot() }),
+  placeholder: styles => ({ ...styles, ...dot() }),
+  singleValue: (styles, { data }) => ({ ...styles, ...dot(data.color) }),
+};
+
+export default () => (
+  <SelectWithValue
+    autoFocus
+    defaultValue={colourOptions[2]}
+    label="Single select"
+    options={colourOptions}
+    styles={colourStylesSingle}
+  />
+);
